+++ CHANGELOG.md
@@ -4,8 +4,8 @@
  - Add support for bounds in FDSN providers.
  - Bugfixes
    - Fixed bug in station_summary.py that was using deprecated _quadrilaterals property of EdgeRupture class.
-
-
+ - Added bandpass and bandstop filter types for StationTrace. Improved modularity of low/high pass filter code.
+ - Added unit tests for the StationTrace filter types.
 
 ## 1.2.4 / 2023-02-09